Update actuatorLineSource.C
Line 544 introduces a new function, windPitching(), which is a modified version of the original harmonicPitching() function. At each time step, it calculates the relative wind speed at the hub height based on the free-stream velocity, the surge velocity of the floating platform, and its pitch angular velocity, and adjusts the blade pitch angle accordingly.
But the specific relative wind speed threshold for variable pitch is directly written in the code and only applies to the IEA 15 MW floating wind turbine. I hope someone interested in this code can help add a function that directly reads the threshold values for each pitch setting from the example folder, thereby expanding its scope of use.